AG-17638 [CLAUDE] Merge caption listener tests into caption.test.ts
Test-only change, at a reviewer's request to keep caption coverage in one file.
`captionEvent.test.ts` is deleted and its suites move into `caption.test.ts` as a
`listeners` block nested inside the existing `describe('Caption')`, so they share
that suite's `setupMockConsole()`, `setupMockCanvas()` and `chart` binding rather
than setting up their own.

- The helpers move inside the new `describe` rather than to module scope, so they
  are not visible to the other caption suites. `options` is renamed
  `chartOptions`, since `caption.test.ts` already uses `options` as a parameter
  name elsewhere.
- The `afterEach` that destroys the chart is scoped to the `listeners` block, so
  the pre-existing suites keep their current teardown behaviour.
- `afterEach`, `vi`, `delay`, `AgCaptionListeners` and `AgCartesianChartOptions`
  are added to the imports; the rest were already present.

No coverage change. The four sub-suites and all 17 listener cases are unchanged
in substance: `caption.test.ts` now runs 62 tests, being the 45 it ran before
(41 declarations, two of them `test.each` over three values) plus the 17 that
report under `Caption > listeners`.
